+++ src/utils/markdown/humanizeAiTextWhitespace.ts
@@ -1,3 +1,5 @@
+import { string_markdown } from "../../types/typeAliases";
+
 /**
  * Change unprintable hard spaces to regular spaces
  *
@@ -5,7 +7,7 @@
  *
  * @public exported from `@promptbook/markdown-utils`
  */
-export function humanizeAiTextWhitespace(aiText: string): string {
+export function humanizeAiTextWhitespace(aiText: string_markdown): string_markdown {
     return aiText.replace(/\u00A0/g, ' ');
 }